Add searching for multiple applications
Summary: When the db search service was written, it was not taken in to account that there could/would be multiple applications depending on the search. As such there is no application filtering. This diff adds per application filtering for bilby and viterbi modules.

Test Plan:
This diff is difficult to test as it requires a mysql server running, however, this repo has extensive tests, and I have added new tests to make sure that per application queries work as expected and results between applications don't bleed.

I think there is low technical risk for this diff as tests still run correctly, including on the CI.

Please have a cursory look at the changes during your review.

Reviewers: mr-asher

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D319