Add ligo job filtering
Summary: A side-effect of having proprietary/public and open/public jobs is that non ligo users should not be able to access the proprietary jobs. This diff simply adds an extra filtering parameter to control if ligo jobs should be included in the search results.

Test Plan: Difficult to test. Test coverage is the most important part of this projects correct functioning. New tests have been added for testing that ligo job filtering works as expected, and they should pass locally.

Reviewers: mr-asher, TomReichardt

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D443