Fix missing filter on end time for uploaded bilby jobs
Summary: This diff adds the missing mysql filter term for the end time for uploaded bilby jobs which was missed in the previous work for filtering on uploaded jobs.

Test Plan: Tests have been added to reflect that filtering works on the end time for uploade jobs as expected. The default list of uploaded jobs should now filter correctly on gwcloud.

Reviewers: TomReichardt, mr-asher, jesmigel

Reviewed By: TomReichardt, jesmigel

Differential Revision: https://phab.adacs.org.au/D729